Make splits persistent in the 'split' file.
This patch adds persistence for the splits. The splits are saved in
the 'split' file on the state directory. Splits are persisted when the
`save-splits` command is called. They are also loaded automatically
loaded at startup, so any splits saved in a previous Helix session can
be recalled in the just started instance.

The patch adds the option `persistence.splits-trim` that specifies the
maximum number of splits entries to have in the splits file.